Added a confirmation when deleting accounts.

+++ src/AccountManagementForm.java
@@ -28,17 +28,21 @@ public class AccountManagementForm {
     confirmButton.addActionListener(new ActionListener() {
       @Override
       public void actionPerformed(ActionEvent e) {
-        String masterPassword = JOptionPane
-            .showInputDialog(null, "Please enter the master password to confirm:", "MBS",
-                JOptionPane.INFORMATION_MESSAGE);
-        if (masterPassword != null && masterPassword.equals("imabeast")) {
-          Account account = System.accountExists(usernameField.getText());
-          if (account != null) {
-            account.setPassword(passwordField.getText());
-            JOptionPane.showMessageDialog(null, "Account password updated!");
-          } else {
-            AccountManagementTool.addAccount(usernameField.getText(), passwordField.getText());
-            JOptionPane.showMessageDialog(null, "Account has been created!");
+        if (usernameField.getText().equals("")) {
+          JOptionPane.showMessageDialog(null, "Account username cannot be empty!");
+        } else {
+          String masterPassword = JOptionPane
+              .showInputDialog(null, "Please enter the master password to confirm:", "MBS",
+                  JOptionPane.INFORMATION_MESSAGE);
+          if (masterPassword != null && masterPassword.equals(System.masterPassword)) {
+            Account account = System.accountExists(usernameField.getText());
+            if (account != null) {
+              account.setPassword(passwordField.getText());
+              JOptionPane.showMessageDialog(null, "Account password updated!");
+            } else {
+              AccountManagementTool.addAccount(usernameField.getText(), passwordField.getText());
+              JOptionPane.showMessageDialog(null, "Account has been created!");
+            }
           }
         }
       }
@@ -60,8 +64,17 @@ public class AccountManagementForm {
                 "Please enter the username you want to delete: " + System.getAccounts(), "MBS",
                 JOptionPane.INFORMATION_MESSAGE);
         if (username != null) {
-          AccountManagementTool.deleteAccount(username);
-          JOptionPane.showMessageDialog(null, "Account " + username + " has been deleted!");
+          if (System.accountExists(username) != null) {
+            String masterPassword = JOptionPane
+                .showInputDialog(null, "Please enter the master password to confirm:", "MBS",
+                    JOptionPane.INFORMATION_MESSAGE);
+            if (masterPassword != null && masterPassword.equals(System.masterPassword)) {
+              AccountManagementTool.deleteAccount(username);
+              JOptionPane.showMessageDialog(null, "Account " + username + " has been deleted!");
+            }
+          } else {
+            JOptionPane.showMessageDialog(null, "Account not found!");
+          }
         }
       }
     });
